+++ earth2mip/initial_conditions/cds.py
@@ -14,9 +14,14 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import warnings
+import eccodes
+from typing import List, Union
 import datetime
+import dataclasses
 from earth2mip import schema
 import xarray
+import numpy as np
 import tempfile
 import os
 from concurrent.futures import ThreadPoolExecutor
@@ -31,183 +36,197 @@ urllib3.disable_warnings(
     urllib3.exceptions.InsecureRequestWarning
 )  # Hack to disable SSL warnings
 
-
-# CDS mapping from channel names to CDS names, tuple means pressure level data
-def create_CDS_channel_mapping():
-    channel_mapping = {
-        "u10m": "10m_u_component_of_wind",
-        "v10m": "10m_v_component_of_wind",
-        "t2m": "2m_temperature",
-        "sp": "surface_pressure",
-        "msl": "mean_sea_level_pressure",
-        "tcwv": "total_column_water_vapour",
-        "u100m": "100m_u_component_of_wind",
-        "v100m": "100m_v_component_of_wind",
-    }
-
-    # for variables depending on pressure level
-    _pressure_levels = [
-        1,
-        2,
-        3,
-        5,
-        7,
-        10,
-        20,
-        30,
-        50,
-        70,
-        100,
-        125,
-        150,
-        175,
-        200,
-        225,
-        250,
-        300,
-        350,
-        400,
-        450,
-        500,
-        550,
-        600,
-        650,
-        700,
-        750,
-        775,
-        800,
-        825,
-        850,
-        875,
-        900,
-        925,
-        950,
-        975,
-        1000,
-    ]
-
-    # add the pressure level variables
-    for level in _pressure_levels:
-        # add u component of wind
-        channel = f"u{level}"
-        cds_name = "u_component_of_wind"
-        channel_mapping[channel] = (cds_name, level)
-
-        # add v component of wind
-        channel = f"v{level}"
-        cds_name = "v_component_of_wind"
-        channel_mapping[channel] = (cds_name, level)
-
-        # add geopotential
-        channel = f"z{level}"
-        cds_name = "geopotential"
-        channel_mapping[channel] = (cds_name, level)
-
-        # add temperature
-        channel = f"t{level}"
-        cds_name = "temperature"
-        channel_mapping[channel] = (cds_name, level)
-
-        # add relative humidity
-        channel = f"r{level}"
-        cds_name = "relative_humidity"
-        channel_mapping[channel] = (cds_name, level)
-    return channel_mapping
+# codes database: https://codes.ecmwf.int/grib/param-db/?filter=grib2
+CHANNEL_TO_CODE = {
+    "z": 129,
+    "u": 131,
+    "v": 132,
+    "t": 130,
+    "q": 133,
+    "r": 157,
+    "t2m": 167,
+    "u10m": 165,
+    "v10m": 166,
+    "u100m": 228246,
+    "v100m": 228247,
+    "tcwv": 137,
+    "sp": 134,
+    "msl": 151,
+    # total precip
+    "tp": 228,
+}
 
 
-def retrieve_channel_data(args):
-    idx, channel, time, area, grid, format, channel_mapping, client = args
-    cds_variable = channel_mapping.get(channel)
-    if not cds_variable:
-        raise ValueError(f"Unknown channel: {channel}")
-
-    if isinstance(cds_variable, tuple):  # it's a pressure level variable
-        cds_name, level = cds_variable
-        tmp_file = tempfile.NamedTemporaryFile(suffix=".nc")
-        client.retrieve(
-            "reanalysis-era5-pressure-levels",
-            {
-                "product_type": "reanalysis",
-                "variable": [cds_name],
-                "pressure_level": [level],
-                "year": time.strftime("%Y"),
-                "month": time.strftime("%m"),
-                "day": time.strftime("%d"),
-                "time": time.strftime("%H:%M"),
-                "area": area,
-                "grid": grid,
-                "format": format,
-            },
-            tmp_file.name,
-        )
-    else:  # it's a single level variable
-        tmp_file = tempfile.NamedTemporaryFile(suffix=".nc")
-        client.retrieve(
-            "reanalysis-era5-single-levels",
-            {
-                "product_type": "reanalysis",
-                "variable": [cds_variable],
-                "year": time.strftime("%Y"),
-                "month": time.strftime("%m"),
-                "day": time.strftime("%d"),
-                "time": time.strftime("%H:%M"),
-                "area": area,
-                "grid": grid,
-                "format": format,
-            },
-            tmp_file.name,
-        )
-    data = xarray.open_dataset(tmp_file.name)
-    main_var = list(set(data.variables) - set(data.coords))[
-        0
-    ]  # main variable is one not in coordinates
-    data = data.rename({main_var: channel})  # rename variable to channel
-    return idx, data
+@dataclasses.dataclass
+class PressureLevelCode:
+    id: int
+    level: int = 0
 
 
-def retrieve_era5_data(
-    channels,
-    time,
-    file_name,
-    area=(90, -180, -90, 180),
-    grid=(0.25, 0.25),
-    format="netcdf",
-):
-    c = Client(progress=False, quiet=True)
-    channel_mapping = create_CDS_channel_mapping()
-    data_arrays = [None] * len(channels)  # pre-allocate list
+@dataclasses.dataclass
+class SingleLevelCode:
+    id: int
 
-    # create a list of arguments for each call to retrieve_channel_data
-    args = [
-        (idx, channel, time, area, grid, format, channel_mapping, c)
-        for idx, channel in enumerate(channels)
-    ]
 
-    with ThreadPoolExecutor() as executor:
-        for idx, data in executor.map(retrieve_channel_data, args):
-            data_arrays[idx] = data  # place data into correct position
+def parse_channel(channel: str) -> Union[PressureLevelCode, SingleLevelCode]:
+    if channel in CHANNEL_TO_CODE:
+        return SingleLevelCode(CHANNEL_TO_CODE[channel])
+    else:
+        code = CHANNEL_TO_CODE[channel[0]]
+        level = int(channel[1:])
+        return PressureLevelCode(code, level=int(level))
 
-    combined_data = xarray.merge(data_arrays)
-    combined_data.to_netcdf(file_name)
+
+@dataclasses.dataclass
+class DataSource:
+    channel_names: List[str]
+    client: Client = dataclasses.field(
+        default_factory=lambda: Client(progress=False, quiet=False)
+    )
+
+    @property
+    def time_means(self):
+        raise NotImplementedError()
+
+    def __getitem__(self, time: datetime.datetime):
+        return _get_channels(self.client, time, self.channel_names)
 
 
 def get(time: datetime.datetime, channel_set: schema.ChannelSet):
-
-    # Get channel data
+    warnings.warn(
+        DeprecationWarning("Will be removed. Please use CDSDataSource instead.")
+    )
     channels = channel_set.list_channels()
+    ds = DataSource(channels)
+    return ds[time]
+
+
+def _get_cds_requests(codes, time, format):
+    grid = (0.25, 0.25)
+    area = (90, -180, -90, 180)
+
+    # create a list of arguments for each call to retrieve_channel_data
+    levels = set()
+    pressure_level_names = set()
+    single_level_names = set()
+    for v in codes:
+        if isinstance(v, PressureLevelCode):  # it's a pressure level variable
+            levels.add(v.level)
+            pressure_level_names.add(v.id)
+        elif isinstance(v, SingleLevelCode):  # it's a single level variable
+            single_level_names.add(v.id)
+
+    if pressure_level_names and levels:
+        yield (
+            "reanalysis-era5-pressure-levels",
+            {
+                "product_type": "reanalysis",
+                "variable": list(pressure_level_names),
+                "pressure_level": sorted(levels),
+                "year": time.strftime("%Y"),
+                "month": time.strftime("%m"),
+                "day": time.strftime("%d"),
+                "time": time.strftime("%H:%M"),
+                "area": area,
+                "grid": grid,
+                "format": format,
+            },
+        )
+
+    if single_level_names:
+        yield (
+            "reanalysis-era5-single-levels",
+            {
+                "product_type": "reanalysis",
+                "variable": sorted(single_level_names),
+                "year": time.strftime("%Y"),
+                "month": time.strftime("%m"),
+                "day": time.strftime("%d"),
+                "time": time.strftime("%H:%M"),
+                "area": area,
+                "grid": grid,
+                "format": format,
+            },
+        )
+
+
+def _parse_files(
+    codes: List[Union[SingleLevelCode, PressureLevelCode]], files: List[str]
+) -> xarray.DataArray:
+    """Retrieve ``codes`` from a list of ``files``
+
+    Returns:
+        a data array of all the codes
+
+    """
+    arrays = [None] * len(codes)
+    for path in files:
+        with open(path) as f:
+            while True:
+                gid = eccodes.codes_grib_new_from_file(f)
+                if gid is None:
+                    break
+                id = eccodes.codes_get(gid, "paramId")
+                level = eccodes.codes_get(gid, "level")
+                type_of_level = eccodes.codes_get(gid, "typeOfLevel")
+
+                if type_of_level == "surface":
+                    code = SingleLevelCode(id)
+                else:
+                    code = PressureLevelCode(id, level=level)
+
+                nlat = eccodes.codes_get(gid, "Nj")
+                nlon = eccodes.codes_get(gid, "Ni")
+
+                lat = eccodes.codes_get_array(gid, "latitudes").reshape(nlat, nlon)
+                lon = eccodes.codes_get_array(gid, "longitudes").reshape(nlat, nlon)
+                vals = eccodes.codes_get_values(gid).reshape(nlat, nlon)
+                eccodes.codes_release(gid)
+
+                try:
+                    i = codes.index(code)
+                except ValueError:
+                    continue
+
+                arrays[i] = vals
+
+    array = np.stack(arrays)
+    coords = {}
+    coords["lat"] = lat[:, 0]
+    coords["lon"] = lon[0, :]
+    return xarray.DataArray(array, dims=["channel", "lat", "lon"], coords=coords)
+
+
+def _download_codes(client, codes, time):
 
     with tempfile.TemporaryDirectory() as d:
-        path = os.path.join(d, "file.nc")
-        retrieve_era5_data(channels, time, path)
-        darray = xarray.open_dataset(path).load()
+        files = []
+        format = "grib"
 
-    # Concatenate channels
-    darray = darray.to_array(dim="channel")
-    darray = darray.transpose("time", "channel", "latitude", "longitude")
+        def download(arg):
+            f = tempfile.mktemp(dir=d, suffix="." + format)
+            name, req = arg
+            path = os.path.join(d, f)
+            client.retrieve(name, req, path)
+            return path
 
-    # Rename lat/lon
-    darray = darray.rename({"latitude": "lat", "longitude": "lon"})
-    darray = darray.assign_coords(lon=darray["lon"] + 180.0)
-    darray = darray.roll(lon=1440 // 2)
+        requests = _get_cds_requests(codes, time, format)
+        with ThreadPoolExecutor(4) as pool:
+            files = pool.map(download, requests)
+
+        darray = _parse_files(codes, files)
 
     return darray
+
+
+def _get_channels(client, time: datetime.datetime, channels: List[str]):
+    codes = [parse_channel(c) for c in channels]
+    darray = _download_codes(client, codes, time)
+    return (
+        darray.assign_coords(channel=channels)
+        .assign_coords(time=time)
+        .expand_dims("time")
+        .transpose("time", "channel", "lat", "lon")
+        .assign_coords(lon=darray["lon"] + 180.0)
+        .roll(lon=1440 // 2)
+    )