+++ CHANGELOG.md
@@ -2,10 +2,15 @@
 
 ## Deprecated
 
+- `earth2mip.initial_conditions.cds.get`
+
 ## Dependencies
+
+- eccodes
 
 ## Features added
 
+- `earth2mip.initial_conditions.cds.CDSDataSource`
 - add `--model-metadata` flag to `earth2mip.inference_medium_range` and
   `earth2mip.inference_ensemble`. Add `metadata` argument to
   `earth2mip.networks.get_model`.